fix: move getStepByMode to seperate utils file
