+++ src/pages/createConfig/createConfig.types.ts
@@ -18,16 +18,3 @@ export type ConfigModeState =
       mode: ConfigFormMode;
     }
   | undefined;
-
-export const getStepByMode = (mode: ConfigFormMode | undefined): StepEnum => {
-  switch (mode) {
-    case 'NEW_CONFIG':
-      return StepEnum.STEP1;
-    case 'NEW_VERSION':
-      return StepEnum.STEP2;
-    case 'ROLLBACK':
-      return StepEnum.STEP3;
-    default:
-      return StepEnum.STEP1;
-  }
-};
